DS-1642 Refactor Graph and WorkflowExecutor's actors
Rationale: logic that should be strictly in Graph class was spread
among other classes (WorkflowExecutorActor, WorkflowNodeExecutorActor).
This made it almost impossible to modify or test (especially writing mocks).

This change:
* splits Graph into two separate classes: DirectedGraph (which is state-agnostic)
but can be sorted topologically and StatefulGraph which is composed of a DirectedGraph
and nodes' states.
* moves all logic connected with Graph's or Nodes' states into StatefulGraph
so that Actors are focused only on the process of execution.
* refactors WorkflowExecutorActor and WorkflowNodeExecutorActor as they contained
state-related logic. Now these classes are much simpler.
* (re)writes tests of the modified classes.

This change also introduces small changes in models module.
However, Json protocol DID NOT change.

Conflicts:
	graph/src/main/scala/io/deepsense/graph/Graph.scala

Change-Id: I05d29fc1f215011290b9ad531863798bb77f3797