DS-2362 Fix 'Execution of Transform fails'
There was a NPE during an execution when a Spark's Estimator
Wrapper was used. It was caused by an incorrect implementation
of replicate. Two fields (the parent estimator and the model) in
the Spark's Estimator Wrapper were set null. Now, the estimator
and the model are replicated too and then put inside the new instance
of the wrapper.

Change-Id: Ia3afe88c387824b8475dfbc5fd2fba6b03a04ffa