Add tests for NotebookRestClient
There has been a lot of development effort targeting the notebooks
integration and the testing has been somewhat neglected. This is to
mitigate that and to give the rest of the team a bit of extra security
when coding some non-trivial changes.

I also included some minor fixes regarding the RestClient and the
polling actor.

Change-Id: Ic4b836a843e0c5f1a667598576c29026ee76373a