Rewrite polling for notebook data
First implementation was a bit of a POC. As it turned out it suffices
for our needs I thought it would be nice to refactor the solution, and
split the idea of implementing polling using akka actor and scheduler
into more general polling mechanism not tied to this specific notebook
rest client use. As a side note - coding this I felt a bit like
reinventing the wheel, maybe there is already some akka-ish/scala-ish
way of doing this coded somewhere?

Tests included.

Change-Id: I58fddf5c22ac49f17923ec0c7247b7a6f8b24235